Start Of Handler
Handler is desinged so that admins can find who place or broke a block, when and how. It should give info on the blocks history and data. This will be stored in an SQLite database with a database per world, a table per chunk and a list of bocks with info in columbs inin the table. A command that allows the admin to punch or place a block in the area they would like to know about allows the admin to retrive saved data. Futher this would run best in a seprate thread as not to affect server preformance.